feat: refactor cache update functions to export them individually
