+++ packages/offix-cache/src/index.ts
@@ -1,4 +1,6 @@
 
-export * from "./cache";
-export * from "./optimisticResponse";
 export * from "./utils/helperFunctions";
+export * from './createMutationOptions';
+export * from "./createOptimisticResponse";
+export * from "./createSubscriptionOptions";
+export * from "./mutationUpdateFunctions";
