--- packages/offix-cache/src/optimisticResponse/index.ts
@@ -1,1 +1,0 @@
-export * from "./optimisticResponse";
